chore: migrates escomplex-module to typescript
