+++ packages/escomplex-module/.eslintrc.json
@@ -1,16 +1,17 @@
 {
-    "env": {
-        "node": true,
-        "es2021": true
-    },
-    "extends": "eslint:recommended",
-    "plugins": ["prettier"],
-    "parserOptions": {
-        "ecmaVersion": 12,
-        "sourceType": "module"
-    },
-    "rules": {
-        "prettier/prettier": "error"
-    }
+  "env": {
+    "node": true,
+    "es2021": true
+  },
+  "parser": "@typescript-eslint/parser",
+  "extends": "eslint:recommended",
+  "plugins": ["prettier", "eslint-plugin-tsdoc"],
+  "parserOptions": {
+    "ecmaVersion": 2018,
+    "sourceType": "module"
+  },
+  "rules": {
+    "prettier/prettier": "error",
+    "tsdoc/syntax": "error"
+  }
 }
-
