+++ packages/escomplex-module/.swcrc
@@ -1,10 +1,16 @@
 {
-    "$schema": "http://json.schemastore.org/swcrc",
-    "module": {
-        "type": "es6",
-        "strict": false,
-        "strictMode": true,
-        "lazy": false,
-        "noInterop": false
+  "$schema": "http://json.schemastore.org/swcrc",
+  "module": {
+    "type": "es6",
+    "strict": false,
+    "strictMode": true,
+    "lazy": false,
+    "noInterop": false
+  },
+  "jsc": {
+    "parser": {
+      "syntax": "typescript",
+      "tsx": false
     }
+  }
 }
