+++ packages/escomplex-module/package.json
@@ -29,8 +29,11 @@
     "@ponticus/babel-parser": "^0.0.1",
     "acorn": "^8.8.0",
     "babylon": "^6.18.0",
+    "@typescript-eslint/eslint-plugin": "^5.42.1",
+    "@typescript-eslint/parser": "^5.42.1",
     "eslint": "^8.26.0",
     "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-tsdoc": "^0.2.17",
     "espree": "^9.4.0",
     "esprima": "^4.0.1"
   },
