+++ packages/escomplex-module/tsconfig.json
@@ -8,5 +8,5 @@
     "outDir": "dist",
     "rootDir": "src"
   },
-  "include": ["src/**/*.js"]
+  "include": ["src/**/*"]
 }
